refactored otlp clients merged into one exporters package
